full quadratic sieve algorithm

+++ 3_RSA/include/TONELLI_SHANKS.h
@@ -7,5 +7,5 @@
 #include  <openssl/rand.h>
 #include "../../1_An_Introduction_To_Cryptography/include/Tools.h"
 ull tonelli_shanks(ull n, ull p);
-ull determine_order_pow(ull t, ull m, ull p);
+int determine_order_pow(ull t, ull m, ull p);
 #endif //CRYPTOGRAPHY_TONELLI_SHANKS_H
